refactor: Criando componente de header
